Fix: Validate game end after the the scores are updated
The condition was at the begging of the playGame(). This causes the game
to intiating extra round to end the game.

Moving the condition after the scores updated makes sure the game ends
immediately after the win/lose scores are achieved and avoid intiating
extra round to trigger the end state